+++ src/postcss-unocss.cjs
@@ -1,0 +1,5 @@
+'use strict';
+
+const mod = require('@unocss/postcss');
+
+module.exports = mod.default || mod;